+++ source/Game/GameObjects/Beton_Wall.h
@@ -51,7 +51,7 @@ public:
 
 	/* Override */
 	virtual void Render() const override;
-	virtual void Update(const uint64_t delta) override;
+	virtual void Update(const double delta) override;
 
 private:
 	void Render_Block(const EBlock_Location block_location) const;
